apps/maia-api/pyproject.toml
[project]
name = "maia-api"
version = "0.1.0"
description = "FastAPI backend for the Ma.ia product."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.14.0",
    "asyncpg>=0.30.0",
    "fastapi>=0.115.0",
    "pydantic-settings>=2.6.0",
    "pyjwt[crypto]>=2.10.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "uvicorn[standard]>=0.30.0",
]

[dependency-groups]
dev = [
    "httpx>=0.27.0",
    "mypy>=1.11.0",
    "pytest>=8.3.0",
    "ruff>=0.6.0",
]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.mypy]
python_version = "3.11"
strict = true
plugins = []

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
